pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tulipa-app-scraper"
version = "0.2.0"
description = "Extracts product and inventory data from the Tulipa B2B florist portal (Helios ERP) to CSV."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Jan Alexandr Kopřiva", email = "jan.alexandr.kopriva@gmail.com" }]
keywords = ["scraper", "helios", "tulipa", "b2b", "czech"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "requests>=2.32.0",
    "beautifulsoup4>=4.13.0",
    "lxml>=5.0.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
]

[project.scripts]
tulipa-scraper = "tulipa_app_scraper.cli:main"

[project.urls]
Homepage = "https://github.com/koprjaa/tulipa-app-scraper"
Repository = "https://github.com/koprjaa/tulipa-app-scraper"
Issues = "https://github.com/koprjaa/tulipa-app-scraper/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/tulipa_app_scraper"]

[tool.ruff]
line-length = 110
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "N",    # pep8-naming
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "N806",  # variable should be lowercase (Helios fields like K1, K2)
    "N818",  # exception name should end in Error (TulipaSessionExpired is a state, not an Error)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-q --strict-markers"
